scraper/requirements.txt
# crawler dependencies
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0

# Data analysis dependencies
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
wordcloud==1.9.2